Physics & Rigidbodies
:)  rigidbodies can be fun.
So learned that the loader re-loads everything that has been added so
for the add object button it reloaded the objects on each click.  Which
made all the rigidbodies go POOF in all directions :)

Assets:
Fixed more mtl files.

App:
* Added input object to loadModel function, instead of multiple
parameters
* Removed global loader because it should not be global.  It reloads all
the tasks so when you added an object it also reloaded the first 7
initially loaded. + Rigidbody chaos!
* Changed camera to arc because it can zoom in & out instead of the free
camera.
* Increased the intensity of lightning since all the models so far seem
to have very dark colors.
* Made the ground global so the addItem can get the BB of it.  Although
for an optimization later the BB itself can be stored as reference
instead.