bounding box detection when placing chairs
It gets a random point and then checks on all other items if it can be
placed there. If it can't it then attempts up to 100 times to do the
same at a new point. It also currently renders points for the new
calculations.

App:
* Removed loading screen when loading a chair, found in firefox as
chrome apperently was a lot faster loading.
* Added debug3d library to draw spheres (points) in the world
* Added Cannon as physics engine to see if there were any differences.
* Added a "setPhysics" function so the loadModels & addItem uses same
function for now.