chairs dont knock each others over anymore!
Technically they still do I think if they simultanously spawn but for
now the physicsImpostor is removed when it collides with the ground.
(Later on grounds)

I'll try and solve the same problem using collision & movement some
other time if needed as for this application that's probably a better
approach. No ragdoll stuff happening then if things go wrong.

A third solution would be to find the nearest ground (which will be
possible later) and lower the object with a fixed speed.