Shadow bugfix
The new ground did not work very well with the directionalLight so it's
a point light currently 20 ypos up.

Added e_item tag to meshes so they can find their item.

Also added bundle red outline for selected item so the whole fan gets
red. Since the mesh isnt a true bundle if you don't pull the string it
will not collide with the beds BB.