Furniture with internal rotation no longer clips through through the surface.
The feature is in a working state but it needs some more work to fix
some bugs. For example the table & drawer rotated through the wall on
other side.

* Added outlineCamera.minZ = camera.minZ
* Added centerOffset, bounding values to calculate centerOffset
* Calculate centerOffset when setting parent. It currently doesn't do
anything on removeChild.
* Added surfaceOffset to cache the offset so it doesn't have to
recalculate often
* Added Delta functions & Absolute updatePosition functions.
updatePositions is for absolute and delta adds to the current Pos!
* Updated adjustCollisionPos to take a vector instead or deltaY value so
it can be used by more functions!
* Updated gravity to send the vector instead of just y value. We can now
in future have furnitures fall towards their parent on drop if needed
* Updated adjustPos to use intersectsMesh( true ) so it goes closer to
bounding boxes
* Added checkMeshCollisions by refactoring the function that used it
before.
* Added getCollisionMeshesForItem so we only need to change this
function for the collisionMeshes.